feat(back-end):move components outside page
