+++ packages/app/pages/index.tsx
@@ -1,82 +1,18 @@
 import type { NextPage } from "next";
 import { useState, useEffect } from "react";
 import { styled } from "@/stitches.config";
-import Card from "@/components/Card";
-import UserCard from "@/components/UserCard";
 import PostList from "@/components/PostList";
-import { useWallet } from "@/components/WalletAuth";
-import { Select } from "@cabindao/topo";
 import { useStore } from "@/store/store";
 import AppState, { Sort } from "@/types";
 import { setupThreadClient, auth } from "@/lib/db";
 import { ThreadID } from "@textile/hub";
 import ClientSide from "@/components/HOC/ClientSide";
+import Title from "@/components/Title";
 import Profile from "@/components/Profile";
+import StickyTabBar from "@/components/TabBar/TabBar";
 
-const TabBarWrapper = styled("div", {
-  boxSizing: "border-box",
-  marginBottom: "$2",
-});
-const TabBarContent = styled("div", {
-  display: "flex",
-  marginBottom: -1,
-  alignItems: "flex-end",
-});
-const TabLink = styled("button", {
-  boxSizing: "border-box",
-  height: "$10",
-  background: "none",
-  paddingLeft: "$4",
-  paddingRight: "$4",
-  border: 0,
-  cursor: "pointer",
-  defaultVariants: {
-    active: false,
-  },
-  variants: {
-    active: {
-      false: {
-        "&:hover": {
-          backgroundColor: "rgba(50, 72, 65, 0.1)",
-        },
-      },
-      true: {
-        backgroundColor: "$forest",
-        color: "$sand",
-      },
-    },
-  },
-});
-const TabButton = (props: any) => <TabLink {...props} />;
-const TabBar = ({ children, ...props }: { children?: React.ReactNode }) => {
-  return (
-    <TabBarWrapper {...props}>
-      <TabBarContent {...props}>{children}</TabBarContent>
-    </TabBarWrapper>
-  );
-};
-const StickyTabBar = styled(TabBar, {
-  backgroundColor: "$sand",
-  borderBottomWidth: 1,
-  borderBottomStyle: "solid",
-  borderColor: "$forest",
-  variants: {
-    position: {
-      fixed: {
-        position: "static",
-      },
-      sticky: {
-        position: "sticky",
-        top: 0,
-        left: 0,
-      },
-    },
-  },
-});
 const Home: NextPage = () => {
-  const { address, ens } = useWallet();
-  const { posts, sort, updateSort } = useStore();
-  const [activeTab, setActiveTab] = useState(0);
+  const { posts, sort } = useStore();
   const [scrolled, setScrolled] = useState(false);
   const handleScroll = () => {
     const offset = window.scrollY;
@@ -86,7 +22,6 @@ const Home: NextPage = () => {
       setScrolled(false);
     }
   };
-
   useEffect(() => {
     window.addEventListener("scroll", handleScroll);
   });
@@ -94,61 +29,7 @@ const Home: NextPage = () => {
     <ClientSide>
       <Profile/>
       <Title>Today</Title>
-      <StickyTabBar position={scrolled ? "sticky" : "fixed"}>
-        {!address && (
-          <TabButton
-            active={activeTab == 0 ? true : false}
-            onClick={() => setActiveTab(0)}
-          >
-            Links
-          </TabButton>
-        )}
-        {address && (
-          <>
-            <TabButton
-              active={activeTab == 0 ? true : false}
-              onClick={() => setActiveTab(0)}
-            >
-              Links
-            </TabButton>
-            <TabButton
-              active={activeTab == 1 ? true : false}
-              onClick={() => setActiveTab(1)}
-            >
-              Submissions
-            </TabButton>
-            <TabButton
-              active={activeTab == 2 ? true : false}
-              onClick={() => setActiveTab(2)}
-            >
-              Upvotes
-            </TabButton>
-            {/* <TabButton
-                active={activeTab == 3 ? true : false}
-                onClick={() => setActiveTab(3)}
-              >
-                Comments
-              </TabButton> */}
-          </>
-        )}
-
-        <div
-          style={{
-            marginLeft: "auto",
-          }}
-        >
-          <Select
-            disabled={false}
-            options={[
-              { key: "newest", label: "Newest" },
-              { key: "trending", label: "Trending" },
-              { key: "controversial", label: "Controversial" },
-            ]}
-            value={"newest"}
-            onChange={(key: Sort) => updateSort(key)}
-          />
-        </div>
-      </StickyTabBar>
+      <StickyTabBar position={scrolled ? "sticky" : "fixed"}/>
       <PostList posts={posts} sort={sort} />
     </ClientSide>
   );
